Change keywords fetch api to use post instead of get
